+++ tools/gen_autotools_acam.pl
@@ -10,11 +10,11 @@
 ##
 ## - $Revision: 1.9 $
 ##
-## - Author: 2025, tomyama
+## - Author: 2025-2026, tomyama
 ## - Intended primarily for personal use, but BSD license permits redistribution.
 ##
 ## BSD 2-Clause License:
-## Copyright (c) 2025, tomyama
+## Copyright (c) 2025-2026, tomyama
 ## All rights reserved.
 ################################################################################
 
@@ -236,11 +236,11 @@ L<GenAutotoolsAcAm_UserFile.pm>, perl(1)
 
 =head1 AUTHOR
 
-2025, tomyama
+2025-2026, tomyama
 
 =head1 LICENSE
 
-Copyright (c) 2025, tomyama
+Copyright (c) 2025-2026, tomyama
 
 All rights reserved.
 
@@ -249,9 +249,11 @@ modification, are permitted provided that the following conditions are met:
 
 1. Redistributions of source code must retain the above copyright notice,
    this list of conditions and the following disclaimer.
+
 2. Redistributions in binary form must reproduce the above copyright notice,
    this list of conditions and the following disclaimer in the documentation
    and/or other materials provided with the distribution.
+
 3. Neither the name of tomyama nor the names of its contributors
    may be used to endorse or promote products derived from this software
    without specific prior written permission.